requirements.txt
# requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
web3==6.11.3
python-dotenv==1.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.9  # For PostgreSQL
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
